changed config and shortcut folder paths
Added a ConfigPath property to the MainPath class, and changed references to this folder in the rest of the code.

Added a ShortcutsPath property to the MainPath class, and changed references to this folder in the rest of the code.

The config and Shortcuts folders are now stored in the user's application data folder.

Started replacing with Path.Combine() the concatenation of paths using slashes and plus signs.